refactor: address PR review comments for consent dialog
